[fix] Fixed implicit conversions from long -> int (#22055)

+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/stats/BrokerStats.java
@@ -20,8 +20,8 @@ package org.apache.pulsar.broker.stats;
 
 public class BrokerStats extends NamespaceStats {
 
-    public int bundleCount;
-    public int topics;
+    public long bundleCount;
+    public long topics;
     public BrokerStats(int ratePeriodInSeconds) {
         super(ratePeriodInSeconds);
     }
